test: isolate fake-timer Wait tests into their own file (Node 26 fix)
On Node 26, jest's useRealTimers() did not cleanly restore the global
setTimeout, so the Wait describe's fake timers leaked and a later test hit
"setTimeout is not defined". Move the Wait tests to test/wait.test.js, which
jest runs in its own sandboxed environment, and drive them entirely with fake
timers (no useRealTimers) so nothing leaks into the rest of the suite.
